webvr/webvr-supported-by-feature-policy.html

<!DOCTYPE html>
<title>Test that vr is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-feature-policy/#dom-featurepolicy-features">
<link rel="help" href="https://immersive-web.github.io/webxr/#feature-policy">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('vr', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise vr.');
</script>
